+++ src/epistasis/dataset.h
@@ -23,9 +23,9 @@ int* get_first_combination_in_block(int order, int block_coordinates[order], int
 
 int get_next_combination_in_block(int order, int comb[order], int block_coordinates[order], int stride);
 
-int **get_genotype_combinations(int order, int *num_combinations);
+uint8_t **get_genotype_combinations(int order, int *num_combinations);
 
-int get_next_genotype_combination(int order, int comb[order]);
+uint8_t get_next_genotype_combination(int order, uint8_t comb[order]);
 
 
 void print_combination(int comb[], unsigned long idx, int k);
